requirements.txt
python-dotenv
flask
openai
gunicorn